+++ src/app/globals.css
@@ -21,12 +21,6 @@ body {
   scrollbar-width: none; /* Firefox */
 }
 
-.highlight {
-  display: inline-block;
-  padding: 3px;
-  border-radius: 3px;
-  background-color: var(--color-primary-200);
-}
 @theme inline {
   --color-background: var(--background);
   --color-foreground: var(--foreground);
@@ -37,6 +31,7 @@ body {
   --color-primary-100: #fff7cf;
   --color-primary-200: #fff3b6;
   --color-primary: #fbe783;
+  --color-primary-bg: #fffdf3;
   --color-secondary: #abe0f0;
   --color-secondary-dark: #71bfd7;
   --color-gray-100: #f0f0f0;
