+++ src/components/history/page.tsx
@@ -1,7 +1,12 @@
+import Header from '../common/Header';
+import FlipCard from './FlipCard';
+
 const HistoryPage = () => {
   return (
-    <div>
-      <h1>상담기록 페이지</h1>
+    <div className="flex flex-col items-center">
+      <Header type="back" />
+      <FlipCard />
+      <p className="script-body-14">카드를 터치해보세요!</p>
     </div>
   );
 };
